Teleferic/requirements.txt
autopep8==1.2.4
base58==0.2.5
pycrypto==2.6.1
coverage==4.4.1
Django==1.11.5
graphene-django==1.3
gunicorn==19.7.1
teamcity-messages==1.21
six==1.10.0
django-behave==0.1.5
behave-teamcity==0.1.25